Use Cython

+++ test.py
@@ -6,7 +6,7 @@ import numpy as np
 from sklearn.datasets import load_digits
 from matplotlib import pyplot
 
-from scw import SCW1
+from scw import SCW1, SCW2
 
 
 def generate_dataset():
@@ -35,7 +35,7 @@ def calc_accuracy(resutls, answers):
 
 X, y = generate_dataset()
 
-N = len(X)*0.8
+N = int(len(X)*0.8)
 training, test = X[:N], X[N:]
 labels, answers = y[:N], y[N:]
